missing trailing comma in make_option tuple
On line 21 of elasticstack.managment.commants.show_document.py there was a missing trailing comma. This caused the "tuple" to be of type instance. Hence trying to add it (+) to BaseCommand.option_list will cause an error. This variable option_list is now similar to to that in show_mapping.py